Correct comparisons for detecting -1 values
-1 values in those fields are set when no ahi.prefs file is present. The
comparisons are meant to avoiding sending these -1 to the driver. The
casts for OutputVolume, Input and Output were causing that code was always
evaluated to false and -1 were passed to the drivers. Code now compares
values based on natural types of fields.

+++ workbench/devs/AHI/Device/device.c
@@ -826,19 +826,19 @@ AllocHardware ( struct AHIDevUnit *iounit,
 
     /* Set hardware properties */
     AHI_ControlAudio(iounit->AudioCtrl,
-        ((LONG)iounit->MonitorVolume == ~0L ? TAG_IGNORE : AHIC_MonitorVolume),
+        (iounit->MonitorVolume == ~0 ? TAG_IGNORE : AHIC_MonitorVolume),
         iounit->MonitorVolume,
 
-        ((LONG)iounit->InputGain == ~0L ? TAG_IGNORE : AHIC_InputGain),
+        (iounit->InputGain == ~0 ? TAG_IGNORE : AHIC_InputGain),
         iounit->InputGain,
 
-        ((ULONG)iounit->OutputVolume == ~0UL ? TAG_IGNORE : AHIC_OutputVolume),
+        (iounit->OutputVolume == ~0 ? TAG_IGNORE : AHIC_OutputVolume),
         iounit->OutputVolume,
 
-        ((ULONG)iounit->Input == ~0UL ? TAG_IGNORE : AHIC_Input),
+        (iounit->Input == ~0UL ? TAG_IGNORE : AHIC_Input),
         iounit->Input,
 
-        ((ULONG)iounit->Output == ~0UL ? TAG_IGNORE : AHIC_Output),
+        (iounit->Output == ~0UL ? TAG_IGNORE : AHIC_Output),
         iounit->Output,
 
         TAG_DONE);
